Upgrade xbk version and add dancing goat store template changes
