+++ Directory.Packages.props
@@ -9,11 +9,11 @@
     <PackageVersion Include="GraphQL" Version="7.8.0" />
     <PackageVersion Include="GraphQL.Client" Version="6.0.2" />
     <PackageVersion Include="GraphQL.Client.Serializer.Newtonsoft" Version="6.0.2" />
-    <PackageVersion Include="Kentico.Xperience.Admin" Version="29.3.2" />
-    <PackageVersion Include="Kentico.Xperience.WebApp" Version="29.3.2" />
-    <PackageVersion Include="kentico.xperience.azurestorage" Version="29.3.2" />
-    <PackageVersion Include="kentico.xperience.imageprocessing" Version="29.3.2" />
-    <PackageVersion Include="Kentico.Xperience.Core" Version="29.3.2" />
+    <PackageVersion Include="Kentico.Xperience.Admin" Version="29.5.3" />
+    <PackageVersion Include="Kentico.Xperience.WebApp" Version="29.5.3" />
+    <PackageVersion Include="kentico.xperience.azurestorage" Version="29.5.3" />
+    <PackageVersion Include="kentico.xperience.imageprocessing" Version="29.5.3" />
+    <PackageVersion Include="Kentico.Xperience.Core" Version="29.5.3" />
     <PackageVersion Include="Moq.AutoMock" Version="3.5.0" />
     <PackageVersion Include="ShopifySharp" Version="6.17.0" />
     <PackageVersion Include="ShopifySharp.Extensions.DependencyInjection" Version="1.6.0" />
