+++ examples/DancingGoat-Shopify/Samples/EmailActivityTracking/EmailActivityTrackingEvaluator.cs
@@ -1,0 +1,49 @@
+﻿using System.Threading.Tasks;
+
+using CMS.ContactManagement;
+using CMS.DataEngine;
+using CMS.DataProtection;
+using CMS.EmailLibrary;
+
+using DancingGoat.Helpers.Generator;
+
+using Kentico.OnlineMarketing.Web.Mvc;
+
+namespace Samples.DancingGoat;
+
+/// <summary>
+/// Example implementation of <see cref="IEmailActivityTrackingEvaluator"/>.
+/// It uses consent generated by the data protection (GDPR) demo. To make it work,
+/// you need to visit Sample data generator application and generate the data.
+/// Without the consent, email activity tracking is disabled.
+/// </summary>
+public class EmailActivityTrackingEvaluator : IEmailActivityTrackingEvaluator
+{
+    private readonly IConsentAgreementService consentAgreementService;
+    private readonly IInfoProvider<ConsentInfo> consentInfoProvider;
+
+
+    /// <summary>
+    /// Initializes an instance of the <see cref="EmailActivityTrackingEvaluator"/> class.
+    /// </summary>
+    public EmailActivityTrackingEvaluator(IConsentAgreementService consentAgreementService, IInfoProvider<ConsentInfo> consentInfoProvider)
+    {
+        this.consentAgreementService = consentAgreementService;
+        this.consentInfoProvider = consentInfoProvider;
+    }
+
+
+    /// <summary>
+    /// Determines whether the email activities can be tracked for the contact.
+    /// </summary>
+    /// <param name="contact">Contact info of the recipient.</param>
+    /// <param name="emailConfiguration">Email configuration info.</param>
+    /// <returns>True if the contact has agreed to tracking by giving consent.</returns>
+    /// <remarks>The consent is generated by <see cref="TrackingConsentGenerator"/>.</remarks>
+    public async Task<bool> IsTrackingAllowed(ContactInfo contact, EmailConfigurationInfo emailConfiguration)
+    {
+        var consent = await consentInfoProvider.GetAsync(TrackingConsentGenerator.CONSENT_NAME);
+
+        return consent is not null && consentAgreementService.IsAgreed(contact, consent);
+    }
+}
